backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
openpyxl==3.1.2
PyPDF2==3.0.1
python-dateutil==2.8.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
Werkzeug==2.3.7
python-dotenv==1.0.0
SQLAlchemy==2.0.20
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
joblib==1.3.2
threadpoolctl==3.2.0
pytz==2023.3
six==1.16.0
psycopg2-binary==2.9.7
supabase==1.0.4
postgrest==0.10.8